fix: use decorator patterns
